Normalized type (again) for ep_greedy

+++ azad/policy.py
@@ -5,6 +5,9 @@ def ep_greedy(x, ep):
     if torch.rand(1) > ep:
         action = torch.argmax(x).unsqueeze(0)
     else:
-        action = torch.randint(0, x.shape[0], (1, ), dtype=torch.int)
+        action = torch.randint(0, x.shape[0], (1, ))
+
+    # Ensure type consistency... 
+    action = action.type(torch.FloatTensor)
 
     return action
